cv-service/main.py
import hashlib
import os
import json
import base64
import numpy as np
from fastapi import FastAPI, UploadFile, File, Form
from fastapi.middleware.cors import CORSMiddleware
import cv2
from PIL import Image
import io
import boto3
from botocore.exceptions import ClientError, NoCredentialsError
from typing import Optional

LIVENESS_THRESHOLD = float(os.environ.get("LIVENESS_THRESHOLD", "0.5"))

# Initialize AWS Bedrock client
try:
    bedrock = boto3.client(
        'bedrock-runtime',
        region_name=os.getenv('AWS_REGION', 'us-east-1'),
        aws_access_key_id=os.getenv('AWS_ACCESS_KEY_ID'),
        aws_secret_access_key=os.getenv('AWS_SECRET_ACCESS_KEY'),
    )
except Exception as e:
    print(f"Warning: Failed to initialize AWS Bedrock client: {e}")
    bedrock = None

app = FastAPI(title="PulsePay CV Service")

# Enable CORS
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)


def get_nova_embedding(image_bytes: bytes) -> list[float]:
    """Get image embedding from Amazon Nova multimodal embeddings via Bedrock."""
    if bedrock is None:
        raise RuntimeError("AWS Bedrock client not initialized. Check AWS credentials.")

    b64_image = base64.b64encode(image_bytes).decode("utf-8")
    body = json.dumps({
        "inputImage": b64_image,
        "embeddingConfig": {
            "outputEmbeddingLength": 1024
        }
    })
    response = bedrock.invoke_model(
        modelId="amazon.nova-embed-multimodal-v1:0",
        body=body,
        contentType="application/json",
        accept="application/json",
    )
    result = json.loads(response['body'].read())
    return result['embedding']


@app.get("/")
async def root():
    return {"status": "PulsePay CV Service running"}


@app.post("/hash-face")
async def hash_face(image: UploadFile = File(...)):
    """
    Extract face embedding using Amazon Nova multimodal embeddings and return SHA3-256 hash.
    """
    try:
        contents = await image.read()

        # Validate image
        nparr = np.frombuffer(contents, np.uint8)
        img = cv2.imdecode(nparr, cv2.IMREAD_COLOR)
        if img is None:
            return {"error": "Invalid image"}

        # Re-encode as JPEG for consistent base64 payload
        success, encoded = cv2.imencode('.jpg', img)
        if not success:
            return {"error": "Failed to encode image"}
        image_bytes = encoded.tobytes()

        try:
            embedding = get_nova_embedding(image_bytes)

            # Hash the embedding with SHA3-256
            embedding_str = ",".join([str(x) for x in embedding])
            hash_obj = hashlib.sha3_256(embedding_str.encode())
            face_hash = hash_obj.hexdigest()

            # Estimate confidence from embedding magnitude (non-zero = valid face signal)
            magnitude = float(np.linalg.norm(embedding))
            confidence = min(99, max(70, int(80 + magnitude * 5)))

            return {
                "hash": face_hash,
                "confidence": confidence,
                "embedding_preview": embedding[:5],
            }

        except (NoCredentialsError, ClientError) as e:
            print(f"AWS Bedrock error: {e}")
            return {
                "error": "AWS credentials missing or invalid. Configure AWS_ACCESS_KEY_ID and AWS_SECRET_ACCESS_KEY.",
            }
        except Exception as e:
            print(f"Nova embedding error: {e}")
            # Fallback to basic image hash
            hash_obj = hashlib.sha3_256(image_bytes)
            return {
                "hash": hash_obj.hexdigest(),
                "confidence": 75,
                "embedding_preview": [],
                "note": "Fallback hash used",
            }

    except Exception as e:
        print(f"Error in hash-face: {e}")
        return {"error": str(e)}


@app.post("/hash-fingerprint")
async def hash_fingerprint(image: UploadFile = File(...)):
    """
    Process fingerprint image and return SHA3-256 hash.
    """
    try:
        contents = await image.read()
        nparr = np.frombuffer(contents, np.uint8)
        img = cv2.imdecode(nparr, cv2.IMREAD_COLOR)

        if img is None:
            return {"error": "Invalid image"}

        # Convert to grayscale
        gray = cv2.cvtColor(img, cv2.COLOR_BGR2GRAY)

        # Apply Gaussian blur
        blurred = cv2.GaussianBlur(gray, (5, 5), 0)

        # Apply binary threshold
        _, binary = cv2.threshold(blurred, 127, 255, cv2.THRESH_BINARY)

        # Find contours
        contours, _ = cv2.findContours(binary, cv2.RETR_TREE, cv2.CHAIN_APPROX_SIMPLE)

        # Compute feature vector from contour areas
        if contours:
            areas = sorted([cv2.contourArea(c) for c in contours if cv2.contourArea(c) > 10])
            feature_vector = areas[:50]
        else:
            feature_vector = cv2.calcHist([gray], [0], None, [256], [0, 256]).flatten().tolist()

        # Convert feature vector to string and hash
        feature_str = ",".join([str(x) for x in feature_vector])
        hash_obj = hashlib.sha3_256(feature_str.encode())
        fingerprint_hash = hash_obj.hexdigest()

        return {
            "hash": fingerprint_hash,
            "features_detected": len(feature_vector),
            "contours_found": len(contours),
        }

    except Exception as e:
        print(f"Error in hash-fingerprint: {e}")
        return {"error": str(e)}


@app.post("/liveness-check")
async def liveness_check(
    image: UploadFile = File(...),
    challenge_type: str = Form(...),
):
    """
    Check if a real face is present (liveness detection) using OpenCV haarcascade
    and optionally verify via Nova embeddings.
    """
    try:
        contents = await image.read()
        nparr = np.frombuffer(contents, np.uint8)
        img = cv2.imdecode(nparr, cv2.IMREAD_COLOR)

        if img is None:
            return {
                "liveness_passed": False,
                "confidence": 0,
                "face_detected": False,
                "error": "Invalid image",
            }

        # Primary face detection with OpenCV haarcascade
        face_cascade = cv2.CascadeClassifier(
            cv2.data.haarcascades + 'haarcascade_frontalface_default.xml'
        )
        gray = cv2.cvtColor(img, cv2.COLOR_BGR2GRAY)
        faces = face_cascade.detectMultiScale(gray, 1.1, 4)

        if len(faces) == 0:
            return {
                "liveness_passed": False,
                "confidence": 0,
                "face_detected": False,
            }

        if len(faces) > 1:
            return {
                "liveness_passed": False,
                "confidence": 50,
                "face_detected": True,
                "faces_count": len(faces),
                "error": "Multiple faces detected",
            }

        # Single face detected — boost confidence with Nova embedding verification
        confidence = 75
        try:
            success, encoded = cv2.imencode('.jpg', img)
            if success:
                embedding = get_nova_embedding(encoded.tobytes())
                magnitude = float(np.linalg.norm(embedding))
                # Higher magnitude indicates stronger face signal
                confidence = min(99, max(70, int(80 + magnitude * 5)))
        except Exception as e:
            print(f"Nova liveness verification skipped: {e}")
            # Proceed with OpenCV-only confidence

        liveness_passed = confidence > (LIVENESS_THRESHOLD * 100)

        return {
            "liveness_passed": liveness_passed,
            "confidence": confidence,
            "face_detected": True,
            "faces_count": 1,
        }

    except Exception as e:
        print(f"Error in liveness-check: {e}")
        return {
            "liveness_passed": False,
            "confidence": 0,
            "face_detected": False,
            "error": str(e),
        }


if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8000)
