cv-service/requirements.txt
fastapi
uvicorn
opencv-python
numpy
Pillow
boto3
python-multipart
psycopg2-binary
